back/requirements.txt
alabaster==0.7.12
amqp==5.0.5
asgiref==3.3.1
Babel==2.9.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.0
Django==3.1.7
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
dnspython==2.1.0
docutils==0.16
drf-yasg==1.20.0
flake8==3.8.4
greenlet==1.0.0
gunicorn==20.0.4
idna==2.9
imagesize==1.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
prompt-toolkit==3.0.16
psycopg2==2.8.6
pyclean==2.0.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
rabbitmq==0.2.0
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
